+++ admin/src/containers/Admin/modules/Creator.tsx
@@ -19,48 +19,68 @@ const Creator: Component = () => {
   return (
     <>
       <div class="admin-panel-editor-header">
-        <span>New Post</span>
+        <h3>New Post</h3>
       </div>
       <div class="admin-panel-editor-form">
-        <label for="title">Title:</label>
+        <label class="admin-panel-editor-form-label" for="title">
+          Title:
+        </label>
         <input
+          class="admin-panel-editor-form-input"
           id="title"
           type="text"
           // onInput={updatePostField('title')}
           value={newPost().title}
         ></input>
-        <label for="slug">Slug:</label>
+        <label class="admin-panel-editor-form-label" for="slug">
+          Slug:
+        </label>
         <input
+          class="admin-panel-editor-form-input"
           id="slug"
           type="text"
           // onInput={updatePostField('title')}
           value={newPost().slug}
         ></input>
-        <label for="published">Published:</label>
+        <div class="admin-panel-editor-form-published">
+          <label class="admin-panel-editor-form-label" for="published">
+            Published:
+          </label>
+          <input
+            class="admin-panel-editor-form-checkbox"
+            id="published"
+            type="checkbox"
+            // onInput={updatePostField('title')}
+            checked={newPost().published}
+          ></input>
+        </div>
+        <label class="admin-panel-editor-form-label" for="date">
+          Date:
+        </label>
         <input
-          id="published"
-          type="checkbox"
-          // onInput={updatePostField('title')}
-          checked={newPost().published}
-        ></input>
-        <label for="date">Date:</label>
-        <input
+          class="admin-panel-editor-form-input"
           id="date"
           type="date"
           // onInput={updatePostField('date')}
           // value={post()?.date.split('T')[0]}
         ></input>
-        <label for="series">Series:</label>
+        <label class="admin-panel-editor-form-label" for="series">
+          Series:
+        </label>
         <input
+          class="admin-panel-editor-form-input"
           type="series"
           // onInput={updatePostField('series')}
           // value={post()?.series}
         ></input>
-        <label for="categories">Categories:</label>
+        <label class="admin-panel-editor-form-label" for="categories">
+          Categories:
+        </label>
         <Index each={newPost().categories}>
           {(c, i) => (
             <>
               <input
+                class="admin-panel-editor-form-input"
                 id="categories"
                 type="text"
                 //   onInput={updatePostField("categories", i)}
@@ -71,22 +91,24 @@ const Creator: Component = () => {
           )}
         </Index>
         {/* <button onClick={addCategory}>+</button> */}
-        <label for="markdown">Markdown:</label>
+        <label class="admin-panel-editor-form-label" for="markdown">
+          Markdown:
+        </label>
         <textarea
-          class="admin-textarea"
+          class="admin-panel-editor-form-textarea"
           id="markdown"
           // onInput={updatePostField('markdown')}
           // value={post()?.markdown}
         ></textarea>
-        {/* <button
-      onClick={async () => {
-        try {
-          await adminUpdatePost(post());
-        } catch (error) {}
-      }}
-    >
-      Update
-    </button> */}
+        <button
+          class="admin-panel-editor-form-button create"
+          // onClick={async () => {
+          //   try {
+          //   } catch (error) {}
+          // }}
+        >
+          Create
+        </button>
       </div>
     </>
   );