+++ admin/src/containers/Admin/modules/PostPanel.tsx
@@ -1,4 +1,4 @@
-import { Component, For } from "solid-js";
+import { Accessor, Component, For } from "solid-js";
 
 import Spinner from "~/components/Spinner/Spinner";
 import { IPost } from "~/api/types";
@@ -6,6 +6,7 @@ import { IPost } from "~/api/types";
 interface IProps {
   posts: IPost[];
   editorUpdatePostSelector: (post: IPost) => (e: any) => void;
+  selectedPost: Accessor<IPost>;
 }
 
 const PostPanel: Component<IProps> = (props) => {
@@ -15,30 +16,43 @@ const PostPanel: Component<IProps> = (props) => {
       <For
         each={props.posts}
         fallback={
-          <div style={{ "margin-top": "15px" }}>
+          <div class="admin-panel-spinner ">
             <Spinner startTime={0}></Spinner>
           </div>
         }
       >
         {(post, i) => (
           <>
-            <div class="admin-panel-post">
+            <div
+              class={`admin-panel-post ${
+                props.selectedPost()?.id === post.id ? "active" : ""
+              }`}
+            >
+              <span class="admin-panel-post-info-title">{post.title}</span>
               <span class="admin-panel-post-info">Slug: {post.slug}</span>
-              <span class="admin-panel-post-info">Title: {post.title}</span>
+              <span class="admin-panel-post-info">
+                Series: {`${post.series}`}
+              </span>
               <span class="admin-panel-post-info">
                 Date: {new Date(post.date).toLocaleDateString()}
+              </span>
+              <span class="admin-panel-post-info">
+                Created: {`${new Date(post.created_at).toLocaleString()}`}
+              </span>
+              <span class="admin-panel-post-info">
+                Updated: {`${new Date(post.updated_at).toLocaleString()}`}
               </span>
               <span class="admin-panel-post-info">
                 Published: {`${post.published}`}
               </span>
               <div>
                 <button
-                  class="admin-panel-post-button"
+                  class="admin-panel-post-button update"
                   onClick={props.editorUpdatePostSelector(post)}
                 >
                   Update
                 </button>
-                <button class="admin-panel-post-button">Delete</button>
+                <button class="admin-panel-post-button delete">Delete</button>
               </div>
             </div>
           </>